Enable Java to use old deprecated type
